Move showMenu() to another class
